refactor: extract shared encoding infrastructure into transcoder::encoding
Move duplicated structs and functions from parallel.rs and
parallel_decode.rs into a shared encoding.rs module:
- EncodingJob, EncodedBlock structs
- encoding_worker, encode_block functions
- send_job_and_drain, buffer_and_write_block, write_single_block
- effective_threads → method on TranscodeConfig

Net -380 lines. No functional changes.
