+++ src/transcoder/mod.rs
@@ -1,5 +1,6 @@
 pub mod block_scanner;
 pub mod boundary;
+mod encoding;
 pub mod parallel;
 pub mod parallel_decode;
 pub mod single;
